Update 50: Started implementing rule matching.
I created the match module which will contain the rule matching
code and the data structures necessary to support it. Currently
I have implemented a stack, a data structure for rules and
morphisms, and a fairly basic static searchplan algorithm.
It's still very rough and does not compile; there are a lot
of changes to be made.