Update 51: Matching Algorithm
Finished coding the logic for a matching algorithm based on a somewhat
unsophisticated static searchplan approach. I have written a function
to populate the morphism data structure once the matching information
has been found. Label matching is still not implemented.

I made some changes to the function headers in the graph module.
The functions now take pointers to nodes and edges instead of
integers. The bodies of the functions are conceptually the same.