Update 56: Debugging
Bugs fixed:

The do-while loop condition in staticSearchplan contained OR instead
of AND.

The node matching loop not break after matching a node; this meant that
more than one node could be matched in a single step which is not the
intended behaviour.

matchEdge did not obtain the correct host node to query for its edge
list.

In the backtracking code, a node mapping was not removed when the node
flag was set to false.