Update 62: Code generation of simple graph matcher.
I have completed the code generation of a simple graph matcher based
on static searchplans and implemented a simple runtime system to
execute the generated code. The 'generate' module is responsible for
the code generation. It performs a depth-first scan of the LHS to
construct a searchplan, and generates code from that searchplan,
emitting a function for each LHS item.

This work necessitated a change to the stack data structure. Instead
of a void pointer, which was causing a rather obscure bug, the stack
data is now represented as a union, which is much cleaner and much
less error-prone.