Update 64: Dangling Condition
I have integrated the checking of the dangling condition into the
generated matching code. The code generation takes an extra argument:
a boolean array of left nodes, where A[i] = true if node i is deleted
by the rule, and uses it to generate the appropriate degree tests on
host nodes in the generated code.

In addition, I have updated the data structure for morphisms: the
StackData now has a struct with two integers representing a
left item -> host item mapping. I also updated stack.c and match.c
to cooperate with this change.