Optimising graph data structure
Wrote a function to return a "good" size for the node pool and
edge pool of a graph from its AST which are passed to the newGraph
function. Used for rule graphs and host graphs.

The runtime data structure used to look up host node indices from
their string IDs (used to add edges to the host graph) has been
changed from a linked list to a rudimentary hash table.