Debugging matching code
Reimplemented the runtime matched_nodes/edges arrays: instead of
storing a bool for each host item, they store a host index for each
rule item. For large host graphs, the former approach was rather
silly!

Also fixed some bugs in the generation of matching code, mostly
related to bidirectional edges. This necessitated the introduction
of bidegrees (the number of bidirectional edges incident to a node)
in the graph data structure.