New stack implementation
Reimplemented stacks as an array of StackData items, in contrast to the
former linked list implementation. restoreGraph takes an extra integer
argument to specify the depth at which the graph to restore is located.
This was done to coordinate with the runtime code which keeps track of
graph backtracking points by recording the current stack depth at points
where backtracking may be needed.