Label Equality
Wrote a function to compare label equality at compile time. This is a
syntactic, distinct from label matching. It is used to determine if
items are relabelled by rules and could also be used as part of the
label matching function.

I have separated generation of the rule matching files from the
generation of the program code generation. This is because the rule
transformation updates the AST with some information to be used in
static analysis prior to program code generation.

There is more character token or character constant type. Characters
are treated exactly as strings of length 1.