Updated semantic checking functions
validateRuleCall and validateProcedureCall have been replaced by
two functions that search the AST for rule declaration nodes and
procedure declaration nodes with a given name. Rule call and
procedure call nodes now point to their associated declarations in
the AST to improve code generation. An added bonus is that checking
for duplicate procedure names no longer requires extensive symbol
table processing: a simple call to the new procedure validation
function suffices. It is not so simple to verify rule declarations
in this way, so I have ignored that for the time being.