Static Analysis
Before code generation there is a static analysis phase that annotates
the AST with various information related to graph copying. If statements,
try statements and loops are annotated with a restore point for
backtracking. These commands, along with rule calls, are annotated with
copy points. This allows for easier code generation as all the hard work
has been done beforehand, so there is no need to hand-code all the special
cases into the code generator, although it's still quite messy in places.

Other big changes:
- Rule and procedure calls in the AST are pointed to their declarations
  in the semantic checking phase. This makes things simpler later on,
  and also allows procedure code to be generated in-line, which removes
  a lot of problems caused by generating a procedure's command sequence
  independently from its context in the main program.
- copyGraph and restoreGraph both take an integer argument: copyGraph
  pushes the copied graph to the passed stack depth, replacing a graph
  if one exists there. restoreGraph pops all graphs on the stack down
  to the depth passed by the caller.
- The code generator is somewhat cleaner: most of the work is now done
  in the rule call generating function, which consequently takes a large
  amount of arguments. Also, correct code is generated for predicate
  rules.