New "graphStacks" module.
Added a new module for the graph backtracking structures and
functions. The graph copying code from the graph module has been
moved to this module. The graph change stack code has been
resurrected, updated with respect to the changes in the graph
data structure, and also moved to the new module.