Reimplemented Static Analysis
The static analysis has been recoded. The previous analysis was
not thorough enough in that it sometimes signalled unnecessary
graph recording. The main improvement made to the analysis is that,
for try conditions and loop bodies, it knows no backtracking is
required if the command sequence cannot fail from the second command
to the last (the first rule call is irrelevant).

Minor Changes and Bug Fixes
- Bidirectional flag and bidegree removed from the host graph data
  structure.
- Reorganised the code generated to evaluate conditions for both
  nodes and variables. This fixed a few bugs and also made the
  generated code a bit nicer.
- Fixed a bug in the rule application that did not print the call
  to add a node in some cases. A related bug is that the node map
  (to get the source and target for added edges) was updated
  before the node was added to the graph.
- Fixed a bug where the last item in the searchplan did not
  remove its match from the morphism on condition failure.
- Refactored the searchplan module.