Host Graph Parsing
Building the host graph at runtime is now done directly from a
parse of the host graph file. This is a huge improvement over the
previous approach of using the parse to generate an AST at compile
time which was then used to generate vast quantities of C code to
build the host graph at runtime. Hence the host graph parser and
lexer have been moved to their own files. The host graph is still
parsed at compile time to get the number of nodes and edges, but
no AST is built. This approach does not yet support the creation
of host graph labels.