New List Data Structure
GP2 lists have been reimplemented as doubly-linked lists so that
they can be built directly from the host graph parser. The modules
most affected by this change are the label module and the genLabel
module. A benefit is that the label matching code is significantly
less clunky than it was previously, at the slight expense of the
implementation of more label management functions.

The main file has been refactored and modified. The GP2-compile
executable now takes different command line arguments than before.
See the README for details.