Optimised assignments in morphisms.
At runtime, variables are identified by a unique integer ID as
opposed to their compile time string identifier. The ID functions
as the index to the assignment of the variable in the morphism.
Because of this, variable lookups can be made in constant time,
in constrast to the previous implementation that iterated over
the assignment array until the assignment with the correct variable
name (string) was located. This made removal of assignments trickier,
so I added an integer stack to the morphism data structure in order
to record the order of assignments made during rule matching.