Host graph ID syntax change
Host graph identifiers can only be integers. The parsers, lexers and
graph printers have been updated to reflect this change. In addition,
the compiler now creates its log in the same directory as the generated
files, and calling 'make build' in the src directory now creates
the shared library in the lib directory.